requirements.txt
ml_metrics
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
dill==0.2.5
docopt==0.6.2
future==0.16.0
imageio==2.4.1
kiwisolver==1.0.1
matplotlib==3.0.0
networkx==2.2
numba==0.40.1
numpy==1.15.2
opencv-contrib-python==3.4.3.18
Pillow==5.3.0
PyFunctional==1.1.3
pyparsing==2.2.2
pyreadline==2.1
python-dateutil==2.7.3
PyWavelets==1.0.1
scikit-image==0.14.0
scipy==1.1.0
six==1.10.0
tabulate==0.7.7
toolz==0.9.0
